Cosmetic fixes for ZOMPLEX Householder reconstruction TESTING routines: 1) In zerrorhr.f zchkorhr.f, "*> \ingroup double_lin" -> "*> \ingroup complex16_lin"; 2) In zchkaa.f, added the missing declaration for ZCHKORHR; 3) In zorhr01.f, changed the type in the description of the argument RESULT from COMPLEX*16 to DOUBLE PECISION;
modified:   TESTING/LIN/zchkaa.f
modified:   TESTING/LIN/zchkorhr.f
modified:   TESTING/LIN/zerrorhr.f
modified:   TESTING/LIN/zorhr01.f

+++ TESTING/LIN/zorhr01.f
@@ -60,7 +60,7 @@
 *>
 *> \param[out] RESULT
 *> \verbatim
-*>          RESULT is COMPLEX*16 array, dimension (6)
+*>          RESULT is DOUBLE PRECISION array, dimension (6)
 *>          Results of each of the six tests below.
 *>          ( C is a M-by-N random matrix, D is a N-by-M random matrix )
 *>
